chore: extract form components from `NewBookScreen`
